Huge commit tidying up trailing whitespace
Makes diffing when your editor removes trailing whitespace harder to see
what's going on for some files

+++ src/inc/Face.h
@@ -15,8 +15,8 @@
 
     You should also have received a copy of the GNU Lesser General Public
     License along with this library in the file named "LICENSE".
-    If not, write to the Free Software Foundation, 51 Franklin Street, 
-    Suite 500, Boston, MA 02110-1335, USA or visit their web page on the 
+    If not, write to the Free Software Foundation, 51 Franklin Street,
+    Suite 500, Boston, MA 02110-1335, USA or visit their web page on the
     internet at http://www.fsf.org/licenses/lgpl.html.
 
 Alternatively, the contents of this file may be used under the terms of the
@@ -170,7 +170,7 @@ class Face::Table
 {
     const Face *            _f;
     mutable const byte *    _p;
-    uint32                  _sz;
+    size_t                  _sz;
     bool                    _compressed;
 
     Error decompress();
